adding customer CRUD

+++ internal/models/customer_model.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Customer struct {
-	CustomerID int16     `json:"customer_id" gorm:"type:integer;primaryKey;autoIncrement"`
+	CustomerID int       `json:"customer_id" gorm:"type:integer;primaryKey;autoIncrement"`
 	StoreID    int16     `json:"store_id" gorm:"type:int2;not null"`
 	FirstName  string    `json:"first_name" gorm:"type:varchar(45);not null"`
 	LastName   string    `json:"last_name" gorm:"type:varchar(45);not null"`
